split observability files to read and payload
